refactor: assets
