+++ config/config.exs
@@ -90,8 +90,8 @@ config :ex_audit,
   ecto_repos: [WraftDoc.Repo],
   version_schema: WraftDoc.ExAudit.Version,
   tracked_schemas: [
+    WraftDoc.Assets.Asset,
     WraftDoc.BlockTemplates.BlockTemplate,
-    WraftDoc.Document.Asset,
     WraftDoc.Document.Block,
     WraftDoc.ContentTypes.ContentType,
     WraftDoc.Document.DataTemplate,
