+++ lib/wraft_doc/document/document.ex
@@ -9,10 +9,11 @@ defmodule WraftDoc.Document do
   alias Ecto.Multi
 
   alias WraftDoc.Account.User
+  alias WraftDoc.Assets
   alias WraftDoc.Client.Minio
   alias WraftDoc.ContentTypes
   alias WraftDoc.ContentTypes.ContentType
-  alias WraftDoc.Document.Asset
+
   alias WraftDoc.Document.Block
   alias WraftDoc.Document.CollectionForm
   alias WraftDoc.Document.CollectionFormField
@@ -23,7 +24,6 @@ defmodule WraftDoc.Document do
   alias WraftDoc.Document.Engine
   alias WraftDoc.Document.Field
   alias WraftDoc.Document.FieldType
-  alias WraftDoc.Document.Frame
   alias WraftDoc.Document.Instance
   alias WraftDoc.Document.Instance.History
   alias WraftDoc.Document.Instance.Version
@@ -1096,99 +1096,6 @@ defmodule WraftDoc.Document do
   def delete_data_template(d_temp), do: Repo.delete(d_temp)
 
   @doc """
-  Create an asset.
-  """
-  # TODO - imprvove tests
-  @spec create_asset(User.t(), map) :: {:ok, Asset.t()}
-  def create_asset(%{current_org_id: org_id} = current_user, params) do
-    params = Map.merge(params, %{"organisation_id" => org_id})
-
-    Multi.new()
-    |> Multi.insert(:asset, current_user |> build_assoc(:assets) |> Asset.changeset(params))
-    |> Multi.update(:asset_file_upload, &Asset.file_changeset(&1.asset, params))
-    |> Repo.transaction()
-    |> case do
-      {:ok, %{asset_file_upload: asset}} -> {:ok, asset}
-      {:error, _, changeset, _} -> {:error, changeset}
-    end
-  end
-
-  def create_asset(_, _), do: {:error, :fake}
-
-  @doc """
-  Index of all assets in an organisation.
-  """
-  # TODO - improve tests
-  @spec asset_index(integer, map) :: map
-  def asset_index(%{current_org_id: organisation_id}, params) do
-    query =
-      from(a in Asset,
-        where: a.organisation_id == ^organisation_id,
-        order_by: [desc: a.inserted_at]
-      )
-
-    Repo.paginate(query, params)
-  end
-
-  def asset_index(_, _), do: {:error, :fake}
-
-  @doc """
-  Show an asset.
-  """
-  # TODO - improve tests
-  @spec show_asset(binary, User.t()) :: %Asset{creator: User.t()}
-  def show_asset(asset_id, user) do
-    with %Asset{} = asset <-
-           get_asset(asset_id, user) do
-      Repo.preload(asset, [:creator])
-    end
-  end
-
-  @doc """
-  Get an asset from its UUID.
-  """
-  # TODO - improve tests
-  @spec get_asset(binary, User.t()) :: Asset.t()
-  def get_asset(<<_::288>> = id, %{current_org_id: org_id}) do
-    case Repo.get_by(Asset, id: id, organisation_id: org_id) do
-      %Asset{} = asset -> asset
-      _ -> {:error, :invalid_id}
-    end
-  end
-
-  def get_asset(<<_::288>>, _), do: {:error, :fake}
-  def get_asset(_, %{current_org_id: _}), do: {:error, :invalid_id}
-
-  @doc """
-  Update an asset.
-  """
-  # TODO - improve tests
-  # file uploading is throwing errors, in tests
-  @spec update_asset(Asset.t(), map) :: {:ok, Asset.t()} | {:error, Ecto.Changset.t()}
-  def update_asset(asset, params) do
-    asset |> Asset.update_changeset(params) |> Repo.update()
-  end
-
-  @doc """
-  Delete an asset.
-  """
-  @spec delete_asset(Asset.t()) :: {:ok, Asset.t()}
-  def delete_asset(asset) do
-    # Delete the uploaded file
-    Repo.delete(asset)
-  end
-
-  @doc """
-  Preload assets of a layout.
-  """
-  @spec preload_asset(Layout.t()) :: Layout.t()
-  def preload_asset(%Layout{} = layout) do
-    Repo.preload(layout, [:assets])
-  end
-
-  def preload_asset(_), do: {:error, :not_sufficient}
-
-  @doc """
   Build a PDF document.
   """
   # TODO  - Write Test
@@ -1210,7 +1117,7 @@ defmodule WraftDoc.Document do
     file_path =
       layout
       |> Repo.preload([:frame])
-      |> download_slug_file()
+      |> Assets.download_slug_file()
 
     System.cmd("cp", ["-a", file_path, base_content_dir])
 
@@ -1243,7 +1150,7 @@ defmodule WraftDoc.Document do
       )
 
     File.write("#{base_content_dir}/content.md", content)
-    pdf_file = pdf_file_path(instance, instance_dir_path, instance_updated?)
+    pdf_file = Assets.pdf_file_path(instance, instance_dir_path, instance_updated?)
 
     pandoc_commands = prepare_pandoc_cmds(pdf_file, base_content_dir)
 
@@ -1252,65 +1159,10 @@ defmodule WraftDoc.Document do
     |> upload_file_and_delete_local_copy(base_content_dir, pdf_file)
   end
 
-  defp download_slug_file(%Layout{frame: nil, slug: slug}),
-    do: :wraft_doc |> :code.priv_dir() |> Path.join("slugs/#{slug}/.")
-
-  defp download_slug_file(%Layout{
-         frame: %Frame{id: frame_id, name: name},
-         organisation_id: organisation_id
-       }) do
-    :wraft_doc
-    |> :code.priv_dir()
-    |> Path.join("slugs/organisation/#{organisation_id}/#{name}/.")
-    |> File.exists?()
-    |> case do
-      true ->
-        :wraft_doc
-        |> :code.priv_dir()
-        |> Path.join("slugs/organisation/#{organisation_id}/#{name}/.")
-
-      false ->
-        slugs_dir =
-          :wraft_doc
-          |> :code.priv_dir()
-          |> Path.join("slugs/organisation/#{organisation_id}/#{name}/.")
-
-        File.mkdir_p!(slugs_dir)
-
-        template_path = Path.join(slugs_dir, "template.tex")
-
-        "organisations/#{organisation_id}/frames/#{frame_id}"
-        |> Minio.download()
-        |> then(&File.write!(template_path, &1))
-
-        slugs_dir
-    end
-  end
-
-  defp pdf_file_path(
-         %Instance{instance_id: instance_id, versions: build_versions},
-         instance_dir_path,
-         true
-       ) do
-    build_versions
-    |> versioned_file_name(instance_id, :next)
-    |> then(&Path.join(instance_dir_path, &1))
-  end
-
-  defp pdf_file_path(
-         %Instance{instance_id: instance_id, versions: build_versions},
-         instance_dir_path,
-         false
-       ) do
-    build_versions
-    |> versioned_file_name(instance_id, :current)
-    |> then(&Path.join(instance_dir_path, &1))
-  end
-
-  defp versioned_file_name(build_versions, instance_id, :current),
+  def versioned_file_name(build_versions, instance_id, :current),
     do: instance_id <> "-v" <> to_string(length(build_versions)) <> ".pdf"
 
-  defp versioned_file_name(build_versions, instance_id, :next),
+  def versioned_file_name(build_versions, instance_id, :next),
     do: instance_id <> "-v" <> to_string(length(build_versions) + 1) <> ".pdf"
 
   defp prepare_markdown(
@@ -1323,7 +1175,7 @@ defmodule WraftDoc.Document do
        ) do
     header =
       Enum.reduce(layout.assets, header, fn x, acc ->
-        find_asset_header_values(x, acc, layout, instance)
+        Assets.find_asset_header_values(x, acc, layout, instance)
       end)
 
     qr_code = Task.await(task)
@@ -1393,30 +1245,6 @@ defmodule WraftDoc.Document do
 
       {_, value} ->
         concat_strings(acc, "#{key}: #{value} \n")
-    end
-  end
-
-  # Find the header values for the content.md file from the assets of the layout used.
-  @spec find_asset_header_values(Asset.t(), String.t(), String.t(), Instance.t()) :: String.t()
-  defp find_asset_header_values(
-         %Asset{name: name, file: file, organisation_id: org_id} = asset,
-         acc,
-         %Layout{frame: frame, slug: slug},
-         %Instance{
-           instance_id: instance_id
-         }
-       ) do
-    binary = Minio.download("organisations/#{org_id}/assets/#{asset.id}/#{file.file_name}")
-
-    asset_file_path =
-      Path.join(File.cwd!(), "organisations/#{org_id}/contents/#{instance_id}/#{file.file_name}")
-
-    File.write!(asset_file_path, binary)
-
-    if frame != nil || slug == "pletter" do
-      concat_strings(acc, "letterhead: #{asset_file_path} \n")
-    else
-      concat_strings(acc, "#{name}: #{asset_file_path} \n")
     end
   end
